attack-tabular/requirements.txt
pandas~=2.0.3
scikit-learn~=1.1.3
datasets~=2.14.5
z3-solver==4.12.4.0
plotly
tqdm
torch~=2.1.1
lightning==2.1.2
numpy~=1.24.3
adversarial-robustness-toolbox==1.16.0

# Training
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
optuna==3.4.0

# Additional ml-model
pytorch-tabnet

# Config
hydra-core==1.3.2